+++ packages/uiengine/src/cli/templates/config.js
@@ -11,13 +11,13 @@ name: ${name}
 # - templates contains the variant preview and application templates
 # - pages is the directory of the UIengine's site structure and page markdown files
 # - data contains sample data that can be referenced in variants and pages
-# - schema contains the optional schema definitions for the components
+# - entities contains the optional entity definitions for the components
 source:
   components: ./src/components
   templates: ./src/templates
   pages: ./${pagesDir}
   data: ./src/uiengine/data
-  schema: ./src/uiengine/schema
+  entities: ./src/uiengine/entities
 
 # Destination paths for the generated output.
 target: ./dist