+++ CHANGELOG.md
@@ -9,3 +9,6 @@
   - The `variantTemplate` entry got added to the config file
   - Templates need to be referred to in page config by full name (including the file extension)
   - Templates are still resolved relative to `source.templates`
+
+- Wording changed from "schema" to "entities" and "properties"
+  - This reflects the actual use case and intention of the feature better